Create a list of single cell experiments

Create a list of single cell experiments from fragment files

getExpListFromFragments(
  fragment.files,
  output.dir = tempdir(),
  tile.size = 500,
  seq.lengths = NULL,
  gene.grs = NULL,
  barcodes.list = NULL,
  BPPARAM = bpparam()
)

Arguments

fragment.files

Vector of strings specifying fragment files. Vector names need to be sample names.

output.dir

String containing the directory where files should be output while creating the MultiAssayExperiment.

tile.size

Integer scalar specifying the size of the tiles in base pairs.

seq.lengths

Named integer vector containing the lengths of the reference sequences used for alignment. Vector names should correspond to the names of the sequences, in the same order of occurrence as in the fragment file. If NULL, this is obtained from the reference genome used by Cellranger (itself located by scanning the header of the fragment file).

gene.grs

Genomic Ranges specifying gene coordinates for creating the gene score matrix. If NA, the gene accessibility matrix will not be created.

barcodes.list

A List with samples as names and the values a vector of barcodes for that sample. If NULL, all barcodes from the fragment file will be used.

BPPARAM

A BiocParallelParam object indicating how matrix creation should be parallelized.

Value

A list of experiments

Author

Natalie Fox